Data Structures. Ruby. BinarySearchTree. Reorganizes to use classes instead of modules.
