--- data-structures/ruby/binary_search_tree/binary_search_tree_node.rb
@@ -1,12 +1,0 @@
-module BinarySearchTree
-
-  def BinarySearchTreeNode(value, parent = nil, left = nil, right = nil)
-    case value
-    when BinarySearchTree::Node
-      value
-    else
-      BinarySearchTree::Node.new(value, parent, left, right)
-    end
-  end
-
-end
